[PATCH] temporary -- return the xyz in the camera frame
Using fillColorPointCloud is too slow. There is an open bug
https://github.com/microsoft/Azure_Kinect_ROS_Driver/issues/56
about it.

Instead, this publish on the xyz values in the camera frame which
is what we need. The nb of points is equal to h*w of the raw_image
and can be easily recombined if needed later on.